requirements.txt
python-dotenv
google-api-python-client
discord.py
